Revert "Move MSSQL helpers to MSSQL project"
This reverts commit 476a207674082906cf5595ac0c0a2fdc326dc8f6.
