+++ Source/EventFlow.MsSql.Tests/IntegrationTests/EventStores/MsSqlEventStoreTests.cs
@@ -25,8 +25,8 @@ using System;
 using System.Threading;
 using EventFlow.MsSql.EventStores;
 using EventFlow.MsSql.Extensions;
-using EventFlow.MsSql.Tests.Helpers;
 using EventFlow.TestHelpers;
+using EventFlow.TestHelpers.MsSql;
 using EventFlow.TestHelpers.Suites;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
