+++ Source/EventFlow.MsSql.Tests/IntegrationTests/EventStores/MsSqlScriptsTests.cs
@@ -24,8 +24,8 @@
 using System.Linq;
 using EventFlow.Extensions;
 using EventFlow.MsSql.EventStores;
-using EventFlow.MsSql.Tests.Helpers;
 using EventFlow.TestHelpers;
+using EventFlow.TestHelpers.MsSql;
 using NUnit.Framework;
 
 namespace EventFlow.MsSql.Tests.IntegrationTests.EventStores
