+++ Source/EventFlow.MsSql.Tests/IntegrationTests/ReadStores/MsSqlReadModelStoreTests.cs
@@ -26,11 +26,11 @@ using System.Threading;
 using EventFlow.Extensions;
 using EventFlow.MsSql.EventStores;
 using EventFlow.MsSql.Extensions;
-using EventFlow.MsSql.Tests.Helpers;
 using EventFlow.MsSql.Tests.IntegrationTests.ReadStores.QueryHandlers;
 using EventFlow.MsSql.Tests.IntegrationTests.ReadStores.ReadModels;
 using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Aggregates.Entities;
+using EventFlow.TestHelpers.MsSql;
 using EventFlow.TestHelpers.Suites;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
