+++ Source/EventFlow.MsSql.Tests/IntegrationTests/ReadStores/ReadModels/MultipleMsSqlDatabasesTests.cs
@@ -31,12 +31,12 @@ using EventFlow.Extensions;
 using EventFlow.MsSql.EventStores;
 using EventFlow.MsSql.Extensions;
 using EventFlow.MsSql.Tests.Extensions;
-using EventFlow.MsSql.Tests.Helpers;
 using EventFlow.ReadStores;
 using EventFlow.Sql.Migrations;
 using EventFlow.Sql.ReadModels.Attributes;
 using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Extensions;
+using EventFlow.TestHelpers.MsSql;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
